Issue #19: Fix mergeConflict.
On official clients, the initial submission of a `LIST_ITEM` node does
NOT have `baseVersion` set. gkeepapi DOES do this, with an initial value
of 1. This change mimics the behavior of the official client and doesn't
send a `baseVersion` till the first update. As a consequence, unsent
nodes now have a `version` of 0.