Added moveToTop
This 'magic' number seems to work for me to indicate that a note should be at the top of the list. If I add a new note in the web interface, I repeatedly get this number for sort, and any value lower than this goes to a random space in the middle of my notes.