Update WebLink to allow missing title and description
Google Assistant's note and list system has been moved to Google Keep:
  https://support.google.com/keep/?p=migrated_from_assistant

These items are annotated with a webLink that looks like this:
  {
    "webLink": {
      "kind": "notes#webLink",
      "url": "https://support.google.com/keep/?p=migrated_from_assistant",
      "provenanceUrl": "https://support.google.com/keep/?p=migrated_from_assistant",
      "hasThumbnailUrl": true
    }
  }

Previously, it was assumed that "name" and "description" are
guaranteed to be supplied in a webLink annotation.  However,
this is no longer the case.  This commit updates the WebLink
class accordingly, and adds a test case.